Setup basic Arc/Verilator cosimulation for Rocket
Create a separate Makefile for the Rocket core that builds an Arcilator
and a Verilator based simulator for the core. Add an abstract model
class with the top-level ports needed by the testbench, and add a
wrapper for the Arcilator and Verilator models, to make them both adhere
to that abstract class. Update the testbench to work against this class.

With these changes, Rocket can be run on Arcilator, Verilator, or both
of them in lockstep. In the lockstep case the testbench aborts if the
simulators start to diverge. Generating simulatenous VCD files is also
supported.