[Boom] Use single clock/reset
Drive boom with a single clock and reset signal. This can be a good test
to check for correctness in the simplified case where there is only a
single clock. Once arcilator generates an eval function (instead of
requiring the user to explicitly call clock functions), we can
reintroduce multiple clocks, or add them as a separate test.